requirements.txt
attrs==25.4.0
beautifulsoup4==4.14.3
bs4==0.0.2
cattrs==26.1.0
certifi==2026.2.25
charset-normalizer==3.4.5
contourpy==1.3.3
cycler==0.12.1
fonttools==4.61.1
idna==3.11
kiwisolver==1.4.9
matplotlib==3.10.8
numpy==2.4.2
packaging==26.0
pandas==3.0.1
pillow==12.1.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
requests==2.32.5
six==1.17.0
soupsieve==2.8.3
SQLAlchemy==2.0.48
tqdm==4.67.3
typing_extensions==4.15.0
urllib3==2.6.3
websocket-client==1.9.0
